Date label and modal

+++ src/app/[locale]/account/profile/saved-searches/SavedSearches.tsx
@@ -9,6 +9,7 @@ import {
     ListItemText,
     Typography,
 } from "@mui/material";
+import dayjs from "dayjs";
 import { flatten } from "lodash";
 import { useTranslations } from "next-intl";
 import useSWRMutation from "swr/mutation";
@@ -17,8 +18,10 @@ import Loading from "@/components/Loading";
 import Paper from "@/components/Paper";
 import useAuth from "@/hooks/useAuth";
 import useGet from "@/hooks/useGet";
+import useModal from "@/hooks/useModal";
 import deleteRequest from "@/services/api/delete";
 import apis from "@/config/apis";
+import { SHORT_DATE_FORMAT } from "@/consts/date";
 import {
     getFiltersFromSaveSearch,
     getUrlFromSearchParams,
@@ -33,6 +36,8 @@ const TRANSLATION_PATH = "pages.saved_searches";
 const SavedSearches = () => {
     const { isLoading } = useAuth();
     const t = useTranslations(TRANSLATION_PATH);
+
+    const { showModal, hideModal } = useModal();
 
     const {
         isLoading: isSavedSearchesLoading,
@@ -51,10 +56,25 @@ const SavedSearches = () => {
             })
     );
 
-    const handleDeleteSavedSearch = async (id: number) => {
-        await deleteSavedSearchesTrigger(id);
+    const handleDeleteSavedSearch = async ({
+        id,
+        name,
+    }: SavedSearchWithPivot) => {
+        showModal({
+            title: t(`deleteModal.title`, { name }),
+            confirmText: t(`deleteModal.confirmButton`),
+            cancelText: t(`deleteModal.cancelButton`),
+            onSuccess: async () => {
+                await deleteSavedSearchesTrigger(id);
 
-        saveSearchesMutate();
+                saveSearchesMutate();
+
+                hideModal();
+            },
+            onCancel: () => {
+                hideModal();
+            },
+        });
     };
 
     if (isLoading || isSavedSearchesLoading) return <Loading />;
@@ -69,96 +89,111 @@ const SavedSearches = () => {
                     </Typography>
                 </Box>
                 <List>
-                    {savedSearchesData?.map(
-                        (
-                            {
-                                search_endpoint,
-                                sort_order,
-                                search_term,
-                                filters,
-                                name,
-                                id,
-                            },
-                            i: number
-                        ) => {
-                            const formattedFilters =
-                                getFiltersFromSaveSearch(filters) || {};
+                    {savedSearchesData?.map((data, i: number) => {
+                        const {
+                            search_endpoint,
+                            sort_order,
+                            search_term,
+                            filters,
+                            name,
+                            updated_at,
+                        } = data;
 
-                            const href = getUrlFromSearchParams(
-                                search_endpoint,
-                                formattedFilters,
-                                sort_order
-                            );
+                        const formattedFilters =
+                            getFiltersFromSaveSearch(filters) || {};
 
-                            const displayFilterValues = flatten(
-                                Object.values(formattedFilters)
-                            ).join(", ");
+                        const href = getUrlFromSearchParams(
+                            search_endpoint,
+                            formattedFilters,
+                            sort_order
+                        );
 
-                            return (
-                                <>
-                                    <ListItem
-                                        secondaryAction={
-                                            <DeleteAction
-                                                id={id}
-                                                onDelete={(id: number) =>
-                                                    handleDeleteSavedSearch(id)
-                                                }
-                                            />
-                                        }
-                                        sx={{ p: 0 }}
-                                        alignItems="flex-start">
-                                        <ListItemButton
-                                            component="a"
-                                            href={href}>
-                                            <ListItemText
-                                                primary={name}
-                                                primaryTypographyProps={{
-                                                    color: "primary",
-                                                    fontWeight: 600,
-                                                    fontSize: 16,
-                                                    mb: 1.5,
-                                                }}
-                                                secondary={
-                                                    <DataList>
+                        const displayFilterValues = flatten(
+                            Object.values(formattedFilters)
+                        ).join(", ");
+
+                        return (
+                            <>
+                                <ListItem
+                                    secondaryAction={
+                                        <DeleteAction
+                                            onDelete={() =>
+                                                handleDeleteSavedSearch(data)
+                                            }
+                                        />
+                                    }
+                                    sx={{ p: 0 }}
+                                    alignItems="flex-start">
+                                    <ListItemButton component="a" href={href}>
+                                        <ListItemText
+                                            primary={name}
+                                            primaryTypographyProps={{
+                                                color: "primary",
+                                                fontWeight: 600,
+                                                fontSize: 16,
+                                                mb: 1.5,
+                                            }}
+                                            secondary={
+                                                <Box
+                                                    sx={{
+                                                        display: "flex",
+                                                        alignItems: "flex-end",
+                                                        pr: 5,
+                                                    }}>
+                                                    <DataList
+                                                        sx={{ flexGrow: 1 }}>
                                                         <DataListItem
-                                                            primary={
-                                                                "Entity Type"
-                                                            }
+                                                            primary={t(
+                                                                "labels.entityType"
+                                                            )}
                                                             secondary={toTitleCase(
                                                                 search_endpoint
                                                             )}
                                                         />
                                                         <DataListItem
-                                                            primary={"Term(s)"}
+                                                            primary={t(
+                                                                "labels.terms"
+                                                            )}
                                                             secondary={
                                                                 search_term ||
-                                                                "None"
+                                                                t("noData")
                                                             }
                                                         />
                                                         <DataListItem
-                                                            primary={
-                                                                "Filter(s)"
-                                                            }
+                                                            primary={t(
+                                                                "labels.filters"
+                                                            )}
                                                             secondary={
                                                                 <Typography color="secondary">
                                                                     {displayFilterValues.length
                                                                         ? displayFilterValues
-                                                                        : "None"}
+                                                                        : t(
+                                                                              "noData"
+                                                                          )}
                                                                 </Typography>
                                                             }
                                                         />
                                                     </DataList>
-                                                }
-                                            />
-                                        </ListItemButton>
-                                    </ListItem>
-                                    {i < savedSearchesData.length - 1 && (
-                                        <Divider />
-                                    )}
-                                </>
-                            );
-                        }
-                    )}
+                                                    <Typography variant="caption">
+                                                        {t("dateSaved", {
+                                                            date: dayjs(
+                                                                updated_at
+                                                            ).format(
+                                                                SHORT_DATE_FORMAT
+                                                            ),
+                                                        })}
+                                                    </Typography>
+                                                </Box>
+                                            }
+                                        />
+                                    </ListItemButton>
+                                </ListItem>
+                                {i < savedSearchesData.length - 1 && (
+                                    <Divider />
+                                )}
+                            </>
+                        );
+                    })}
                 </List>
             </Box>
         </Paper>
